Move PluginManager to plugin_manager.py
